old-tests/submission/Opera/preload/none/preload-none-to-invalid-event-order-ogg.html

<!doctype html>
<!-- This file was generated from 'preload-none-to-invalid-event-order.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:none event order when preload is set to invalid after suspend is fired - ogg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="none" controls></audio>
<video preload="none" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:none event order when preload is set to invalid after suspend is fired - ogg", "HTMLVideoElement.preload:none event order when preload is set to invalid after suspend is fired - ogg", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
	var tag_name = vars[1];
	var events_actual = '';
	var node = document.getElementsByTagName(tag_name)[0];
	var skip_once = true; // start_event = end_event = suspend, so endTest() should be called on the second occurence of the event
	var events  = ['loadstart', 'progress', 'suspend', 'abort', 'error', 'loadedmetadata', 'loadeddata', 'canplay', 'canplaythrough'];
	for (var e in events) {
	    node.addEventListener(events[e], handleEvent, false);
	}
	node.addEventListener('suspend', startTest, false);
	node.addEventListener('error', function() {
	    if (!node.canPlayType(tag_name + '/ogg')) {
		t.step(function() {
		    assert_unreached("Cannot play '" + tag_name + "/ogg'");
		    t.done();
		});
	    }
	}, false);
	node.src = '../resources/preload.ogv?pipe=trickle(100000:d1:r2)&nocache=' + Math.random();

	function startTest(e) {
	    node.removeEventListener('suspend', startTest, false);
	    node.preload = 'invalid';
	}

	function handleEvent(e) {
	    events_actual += e.type + ' ';
	    if (e.type == 'suspend') {
		if (!skip_once) endTest();
		else skip_once = false;
	    }
	}

	function endTest() {
	    t.step(function() {
		assert_regexp_match(events_actual, /^loadstart suspend (progress )+loadedmetadata (progress )*loadeddata (progress )*suspend $/g, 'Event order check');
		assert_equals(node.preload, 'metadata', "'preload' value should be 'metadata'");
		t.done();
	    });
	}
    });
});
</script>
